change: 重构 query 功能
