--- seal/db/base_query.py
@@ -1,5 +1,0 @@
-import abc
-
-
-class BaseQuery(metaclass=abc.ABCMeta):
-    pass
